requirements.txt
chardet>=4.0.0
Pillow>=8.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
pygments>=2.15.0 